Remove the separation between `user` and `reserved_email` packages.
