+++ model/src/main/proto/spine_examples/chatspn/account/reserved_email_events.proto
@@ -25,13 +25,13 @@
  */
 syntax = "proto3";
 
-package spine_examples.chatspn.account.reserved_email;
+package spine_examples.chatspn.account;
 
 import "spine/options.proto";
 
 option (type_url_prefix) = "type.chatspn.spine.io";
-option java_package = "io.spine.examples.chatspn.account.reservedemail.event";
-option java_outer_classname = "EventsProto";
+option java_package = "io.spine.examples.chatspn.account.event";
+option java_outer_classname = "ReservedEmailEventsProto";
 option java_multiple_files = true;
 
 import "spine/core/user_id.proto";
