+++ model/src/main/proto/spine_examples/chatspn/account/user_commands.proto
@@ -25,13 +25,13 @@
  */
 syntax = "proto3";
 
-package spine_examples.chatspn.account.user;
+package spine_examples.chatspn.account;
 
 import "spine/options.proto";
 
 option (type_url_prefix) = "type.chatspn.spine.io";
-option java_package = "io.spine.examples.chatspn.account.user.command";
-option java_outer_classname = "CommandsProto";
+option java_package = "io.spine.examples.chatspn.account.command";
+option java_outer_classname = "UserCommandsProto";
 option java_multiple_files = true;
 
 import "spine/core/user_id.proto";
