+++ model/src/main/proto/spine_examples/chatspn/account/user_profile.proto
@@ -23,14 +23,30 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
+syntax = "proto3";
 
-/**
- * Provides data types that describe events with reserved email.
- */
-@CheckReturnValue
-@ParametersAreNonnullByDefault
-package io.spine.examples.chatspn.account.reservedemail.event;
+package spine_examples.chatspn.account;
 
-import com.google.errorprone.annotations.CheckReturnValue;
+import "spine/options.proto";
 
-import javax.annotation.ParametersAreNonnullByDefault;
+option (type_url_prefix) = "type.chatspn.spine.io";
+option java_package = "io.spine.examples.chatspn.account";
+option java_outer_classname = "UserProfileProto";
+option java_multiple_files = true;
+
+import "google/protobuf/timestamp.proto";
+import "spine/core/user_id.proto";
+
+// A user profile view in the chatting system.
+message UserProfile {
+    option (entity) = { kind: PROJECTION };
+
+    // The ID of the user.
+    spine.core.UserId id = 1;
+
+    // A name of a user.
+    string name = 2 [(required) = true];
+
+    // A time of last user activity.
+    google.protobuf.Timestamp when_last_active = 3;
+}
