+++ server/src/main/java/io/spine/examples/chatspn/server/ChatsContext.java
@@ -26,13 +26,13 @@
 
 package io.spine.examples.chatspn.server;
 
+import io.spine.examples.chatspn.server.account.reservedemail.ReservedEmailAggregate;
+import io.spine.examples.chatspn.server.account.user.UserAggregate;
+import io.spine.examples.chatspn.server.account.user.UserProfileRepository;
 import io.spine.examples.chatspn.server.chat.ChatAggregate;
 import io.spine.examples.chatspn.server.chat.ChatMembersRepository;
 import io.spine.examples.chatspn.server.message.MessageAggregate;
-import io.spine.examples.chatspn.server.account.reservedemail.ReservedEmailAggregate;
 import io.spine.examples.chatspn.server.message.MessageSendingRepository;
-import io.spine.examples.chatspn.server.account.user.UserAggregate;
-import io.spine.examples.chatspn.server.account.user.UserProfileRepository;
 import io.spine.server.BoundedContext;
 import io.spine.server.BoundedContextBuilder;
 import io.spine.server.DefaultRepository;
